+++ tests/expectations/tests/class.rs
@@ -335,6 +335,7 @@ impl Default for IncompleteArrayNonCopiable {
     }
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union Union {
     pub d: f32,