+++ tests/expectations/tests/layout_mbuf.rs
@@ -127,6 +127,7 @@ fn bindgen_test_layout_rte_atomic16_t() {
 }
 /// The generic rte_mbuf, containing a packet mbuf.
 #[repr(C)]
+#[repr(align(64))]
 pub struct rte_mbuf {
     pub cacheline0: MARKER,
     ///< Virtual address of segment buffer.
@@ -179,6 +180,7 @@ pub struct rte_mbuf {
 /// or non-atomic) is controlled by the CONFIG_RTE_MBUF_REFCNT_ATOMIC
 /// config option.
 #[repr(C)]
+#[repr(align(2))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_1 {
     ///< Atomically accessed refcnt
@@ -228,6 +230,7 @@ impl Default for rte_mbuf__bindgen_ty_1 {
     }
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_2 {
     ///< L2/L3/L4 and tunnel information.
@@ -236,10 +239,10 @@ pub union rte_mbuf__bindgen_ty_2 {
     _bindgen_union_align: u32,
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Debug, Default, Copy, Clone, Hash, PartialEq, Eq)]
 pub struct rte_mbuf__bindgen_ty_2__bindgen_ty_1 {
     pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4usize], u8>,
-    pub __bindgen_align: [u32; 0usize],
 }
 #[test]
 fn bindgen_test_layout_rte_mbuf__bindgen_ty_2__bindgen_ty_1() {
@@ -412,6 +415,7 @@ impl Default for rte_mbuf__bindgen_ty_2 {
     }
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_3 {
     ///< RSS hash result if RSS enabled
@@ -431,6 +435,7 @@ pub struct rte_mbuf__bindgen_ty_3__bindgen_ty_1 {
     pub hi: u32,
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_3__bindgen_ty_1__bindgen_ty_1 {
     pub __bindgen_anon_1: rte_mbuf__bindgen_ty_3__bindgen_ty_1__bindgen_ty_1__bindgen_ty_1,
@@ -667,6 +672,7 @@ impl Default for rte_mbuf__bindgen_ty_3 {
     }
 }
 #[repr(C)]
+#[repr(align(8))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_4 {
     ///< Can be used for external metadata
@@ -714,6 +720,7 @@ impl Default for rte_mbuf__bindgen_ty_4 {
     }
 }
 #[repr(C)]
+#[repr(align(8))]
 #[derive(Copy, Clone)]
 pub union rte_mbuf__bindgen_ty_5 {
     ///< combined for easy fetch
@@ -722,10 +729,10 @@ pub union rte_mbuf__bindgen_ty_5 {
     _bindgen_union_align: u64,
 }
 #[repr(C)]
+#[repr(align(8))]
 #[derive(Debug, Default, Copy, Clone, Hash, PartialEq, Eq)]
 pub struct rte_mbuf__bindgen_ty_5__bindgen_ty_1 {
     pub _bitfield_1: __BindgenBitfieldUnit<[u8; 8usize], u16>,
-    pub __bindgen_align: [u64; 0usize],
 }
 #[test]
 fn bindgen_test_layout_rte_mbuf__bindgen_ty_5__bindgen_ty_1() {
@@ -887,6 +894,11 @@ fn bindgen_test_layout_rte_mbuf() {
         ::std::mem::size_of::<rte_mbuf>(),
         128usize,
         concat!("Size of: ", stringify!(rte_mbuf))
+    );
+    assert_eq!(
+        ::std::mem::align_of::<rte_mbuf>(),
+        64usize,
+        concat!("Alignment of ", stringify!(rte_mbuf))
     );
     assert_eq!(
         unsafe { &(*(::std::ptr::null::<rte_mbuf>())).cacheline0 as *const _ as usize },