+++ tests/expectations/tests/libclang-3.8/error-E0600-cannot-apply-unary-negation-to-u32.rs
@@ -6,5 +6,6 @@
     non_camel_case_types,
     non_upper_case_globals
 )]
+#![allow(overflowing_literals)]
 
 pub const a: u32 = 18446744073709551611;