+++ tests/expectations/tests/union_bitfield.rs
@@ -90,6 +90,7 @@ where
     }
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union U4 {
     pub _bitfield_1: __BindgenBitfieldUnit<[u8; 1usize], u8>,
@@ -137,6 +138,7 @@ impl U4 {
     }
 }
 #[repr(C)]
+#[repr(align(4))]
 #[derive(Copy, Clone)]
 pub union B {
     pub _bitfield_1: __BindgenBitfieldUnit<[u8; 4usize], u32>,